*: move joinversion from executor/join/ to util/ to break planner→executor dependency
The joinversion package has no executor-specific dependencies (only uses
strings and unsafe) but was located under pkg/executor/join/, creating a
false dependency from planner, sessionctx, and distsql into executor
internals. Move it to pkg/util/joinversion/ as a shared utility.
