+++ pkg/distsql/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "//pkg/ddl/placement",
         "//pkg/distsql/context",
         "//pkg/errctx",
-        "//pkg/executor/join/joinversion",
         "//pkg/expression",
         "//pkg/infoschema/context",
         "//pkg/kv",
@@ -34,6 +33,7 @@ go_library(
         "//pkg/util/dbterror",
         "//pkg/util/execdetails",
         "//pkg/util/intest",
+        "//pkg/util/joinversion",
         "//pkg/util/logutil",
         "//pkg/util/memory",
         "//pkg/util/ranger",
