+++ pkg/distsql/context/BUILD.bazel
@@ -30,12 +30,12 @@ go_test(
     flaky = True,
     deps = [
         "//pkg/errctx",
-        "//pkg/executor/join/joinversion",
         "//pkg/kv",
         "//pkg/parser/mysql",
         "//pkg/util/context",
         "//pkg/util/deeptest",
         "//pkg/util/execdetails",
+        "//pkg/util/joinversion",
         "//pkg/util/memory",
         "//pkg/util/ppcpuusage",
         "//pkg/util/sqlkiller",
