+++ pkg/executor/BUILD.bazel
@@ -136,7 +136,6 @@ go_library(
         "//pkg/executor/internal/util",
         "//pkg/executor/internal/vecgroupchecker",
         "//pkg/executor/join",
-        "//pkg/executor/join/joinversion",
         "//pkg/executor/lockstats",
         "//pkg/executor/metrics",
         "//pkg/executor/sortexec",
@@ -243,6 +242,7 @@ go_library(
         "//pkg/util/hack",
         "//pkg/util/hint",
         "//pkg/util/intest",
+        "//pkg/util/joinversion",
         "//pkg/util/keydecoder",
         "//pkg/util/logutil",
         "//pkg/util/logutil/consistency",
