+++ pkg/executor/join/BUILD.bazel
@@ -38,7 +38,6 @@ go_library(
         "//pkg/executor/internal/exec",
         "//pkg/executor/internal/testutil",
         "//pkg/executor/internal/vecgroupchecker",
-        "//pkg/executor/join/joinversion",
         "//pkg/executor/unionexec",
         "//pkg/expression",
         "//pkg/parser/mysql",
@@ -60,6 +59,7 @@ go_library(
         "//pkg/util/execdetails",
         "//pkg/util/hack",
         "//pkg/util/intest",
+        "//pkg/util/joinversion",
         "//pkg/util/logutil",
         "//pkg/util/memory",
         "//pkg/util/mvmap",
