+++ pkg/planner/core/BUILD.bazel
@@ -85,7 +85,6 @@ go_library(
         "//pkg/config/kerneltype",
         "//pkg/domain",
         "//pkg/errctx",
-        "//pkg/executor/join/joinversion",
         "//pkg/expression",
         "//pkg/expression/aggregation",
         "//pkg/expression/exprctx",
@@ -170,6 +169,7 @@ go_library(
         "//pkg/util/hint",
         "//pkg/util/intest",
         "//pkg/util/intset",
+        "//pkg/util/joinversion",
         "//pkg/util/kvcache",
         "//pkg/util/logutil",
         "//pkg/util/memory",
@@ -251,7 +251,6 @@ go_test(
         "//pkg/config",
         "//pkg/config/kerneltype",
         "//pkg/domain",
-        "//pkg/executor/join/joinversion",
         "//pkg/expression",
         "//pkg/expression/aggregation",
         "//pkg/expression/exprctx",
@@ -304,6 +303,7 @@ go_test(
         "//pkg/util/dbterror",
         "//pkg/util/dbterror/plannererrors",
         "//pkg/util/hint",
+        "//pkg/util/joinversion",
         "//pkg/util/logutil",
         "//pkg/util/mock",
         "//pkg/util/plancodec",
