+++ pkg/sessionctx/variable/BUILD.bazel
@@ -23,7 +23,6 @@ go_library(
         "//pkg/config",
         "//pkg/config/kerneltype",
         "//pkg/errno",
-        "//pkg/executor/join/joinversion",
         "//pkg/infoschema/issyncer/mdldef",
         "//pkg/keyspace",
         "//pkg/kv",
@@ -55,6 +54,7 @@ go_library(
         "//pkg/util/execdetails",
         "//pkg/util/gctuner",
         "//pkg/util/intest",
+        "//pkg/util/joinversion",
         "//pkg/util/kvcache",
         "//pkg/util/logutil",
         "//pkg/util/mathutil",
@@ -113,13 +113,13 @@ go_test(
     deps = [
         "//pkg/config",
         "//pkg/config/kerneltype",
-        "//pkg/executor/join/joinversion",
         "//pkg/kv",
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
         "//pkg/sessionctx/vardef",
         "//pkg/testkit/testsetup",
         "//pkg/util/gctuner",
+        "//pkg/util/joinversion",
         "//pkg/util/memory",
         "//pkg/util/timeutil",
         "//pkg/util/traceevent",
