+++ pkg/util/joinversion/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "joinversion",
     srcs = ["join_version.go"],
-    importpath = "github.com/pingcap/tidb/pkg/executor/join/joinversion",
+    importpath = "github.com/pingcap/tidb/pkg/util/joinversion",
     visibility = ["//visibility:public"],
 )
